Support 2 billion atoms.

+++ psfgen/src/psf_file_extract.c
@@ -569,7 +569,7 @@ int psf_file_extract(topo_mol *mol, FILE *file, FILE *pdbfile, FILE *namdbinfile
     char tmpc;
     char static_assert_int_is_32_bits[sizeof(int) == 4 ? 1 : -1];
     if ( ! atomcoords ) {
-      atomcoords = (double *)malloc(natoms * 3 * sizeof(double));
+      atomcoords = (double *)malloc(natoms * 3L * sizeof(double));
     }
     fseek(namdbinfile,0,SEEK_END);
     numatoms = (ftell(namdbinfile)-4)/24;
@@ -610,8 +610,8 @@ int psf_file_extract(topo_mol *mol, FILE *file, FILE *pdbfile, FILE *namdbinfile
     if (wrongendian) {
       print_msg(v,"namdbin file appears to be other-endian");
     }
-    if (fread(atomcoords, sizeof(double), 3 * natoms, namdbinfile)
-                                 != (size_t)(3 * natoms)) {
+    if (fread(atomcoords, sizeof(double), 3L * natoms, namdbinfile)
+                                 != (size_t)(3L * natoms)) {
       print_msg(v,"error reading data from namdbin file");
       free(atomlist);
       free(atomcoords);
@@ -680,9 +680,9 @@ int psf_file_extract(topo_mol *mol, FILE *file, FILE *pdbfile, FILE *namdbinfile
     if (wrongendian) {
       print_msg(v,"namdbin file appears to be other-endian");
     }
-    atomvels = (double *)malloc(natoms * 3 * sizeof(double));
-    if (fread(atomvels, sizeof(double), 3 * natoms, velnamdbinfile)
-                                 != (size_t)(3 * natoms)) {
+    atomvels = (double *)malloc(natoms * 3L * sizeof(double));
+    if (fread(atomvels, sizeof(double), 3L * natoms, velnamdbinfile)
+                                 != (size_t)(3L * natoms)) {
       print_msg(v,"error reading data from namdbin file");
       free(atomlist);
       free(atomcoords);
@@ -690,11 +690,11 @@ int psf_file_extract(topo_mol *mol, FILE *file, FILE *pdbfile, FILE *namdbinfile
       return -1;
     }
     if (wrongendian) {
-      int i;
+      long i;
       char tmp0, tmp1, tmp2, tmp3;
       char *cdata = (char *) atomcoords;
       print_msg(v,"converting other-endian data from namdbin file");
-      for ( i=0; i<3*natoms; ++i, cdata+=8 ) {
+      for ( i=0; i<3L*natoms; ++i, cdata+=8 ) {
         tmp0 = cdata[0]; tmp1 = cdata[1];
         tmp2 = cdata[2]; tmp3 = cdata[3];
         cdata[0] = cdata[7]; cdata[1] = cdata[6];
